chore: pin Rust toolchain version
Create a file that pins the used Rust toolchain to a specific version.
This isn't super useful as this is a library crate and consumers won't
see any effect from this file, but it will prevent the examples from
breaking due to changes in newer versions. It also helps prevent
warnings when checking out older commits, as the toolchain version is
now pinned in every commit.